% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roll.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle}
\usage{
shuffle(cards)
}
\arguments{
\item{Regular}{USA plying cards}
}
\value{
randome card
}
\description{
Shuffle
}
\examples{
cards <- readr::read_csv("https://gist.githubusercontent.com/garrettgman/9629323/raw/ee5dfc039fd581cb467cc69c226ea2524913c3d8/deck.csv")
deck2 <- shuffle(cards = deck)
deck2[1:5, ]
}

